+++ test/web/plugs/federating_plug_test.exs
@@ -4,7 +4,8 @@
 
 defmodule Pleroma.Web.FederatingPlugTest do
   use Pleroma.Web.ConnCase
-  clear_config_all([:instance, :federating])
+
+  clear_config([:instance, :federating])
 
   test "returns and halt the conn when federating is disabled" do
     Pleroma.Config.put([:instance, :federating], false)